Remove crossOrigin flag in form-smippet sync()
When we initially built this function, the sync() call hit an endpoint that was
HTTPS, and the page was served via HTTP.  This meant the request was CORS.
We changed the endpoint to be the same protocol as the page, but we left
the crossOrigin flag in the options for the sync() function.  This didn't
mean anything for most browsers (Chrome, FF, IE10+) but IE9 and below
force CORS requests to use a special request object which is not
supported by Backbone. So this flag essentially broke this feature in IE9 and
below.

This drops the flag entirely from the options for the sync() function,
reverting to the default.  This fixes the form builder in IE9, and has no
apparent effect in other browsers.

I tested this on sites that have ssl-everywhere disabled as well as those that
have ssl-everywhere enabled.